+++ VoiceReminder/Moduls/ListVoiceModul/CoreData/Voice+CoreDataClass.swift
@@ -1,0 +1,15 @@
+//
+//  Voice+CoreDataClass.swift
+//  VoiceReminder
+//
+//  Created by Daniil Litvinov on 10.02.2023.
+//
+//
+
+import Foundation
+import CoreData
+
+@objc(Voice)
+public class Voice: NSManagedObject {
+
+}
