+++ VoiceReminder/Moduls/MainScreenModul/ModelView/MainViewModel.swift
@@ -1,0 +1,8 @@
+//
+//  MainViewModel.swift
+//  VoiceReminder
+//
+//  Created by Daniil Litvinov on 10.02.2023.
+//
+
+import Foundation
