+++ VoiceReminder/Resources/Info.plist
