+++ VoiceReminder/Resources/VoiceReminder.xcdatamodeld/VoiceReminder.xcdatamodel/contents
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="21513" systemVersion="22C65" minimumToolsVersion="Automatic" sourceLanguage="Swift" userDefinedModelVersionIdentifier="">
+    <entity name="Voice" representedClassName="Voice" syncable="YES">
+        <attribute name="voiceAT" optional="YES" attributeType="Binary"/>
+    </entity>
+</model>
